finalized world saves and cleaned up code. should be the final commit for a while.

+++ src/World/World.h
@@ -38,7 +38,7 @@ private:
 	ChunkManager m_Manager;
 	Player m_Player;
 
-	int m_LastBlockQueueSize;
+	size_t m_LastBlockQueueSize;
 	bool m_ResetBuildVars;
 	bool m_ShouldCloseChunkerThread;
